+++ packages/hooks/src/use-data-source-status.ts
@@ -1,5 +1,4 @@
 import type { Status, StatusAdapter } from '@affine/y-provider';
-import { noop } from 'foxact/noop';
 import { useCallback, useSyncExternalStore } from 'react';
 
 type UIStatus =
@@ -10,24 +9,7 @@ type UIStatus =
 
 export function useDataSourceStatus(datasource: StatusAdapter): UIStatus {
   return useSyncExternalStore(
-    useCallback(
-      onStoreChange => {
-        if (datasource.subscribeStatusChange) {
-          return datasource.subscribeStatusChange(onStoreChange);
-        } else {
-          return noop;
-        }
-      },
-      [datasource]
-    ),
-    useCallback(() => {
-      if (datasource.getStatus) {
-        return datasource.getStatus();
-      } else {
-        return {
-          type: 'unknown',
-        };
-      }
-    }, [datasource])
+    datasource.subscribeStatusChange,
+    useCallback(() => datasource.status, [datasource])
   );
 }
